add update db in routes

+++ interet_web_site/app/routes.py
@@ -1,6 +1,7 @@
 from flask import Blueprint, render_template, request, jsonify, current_app
 from .models import db, Device, Credential, File, Network
 import base64, os, uuid
+import services
 
 main = Blueprint("main", __name__)
 
@@ -18,42 +19,72 @@ def device(device_id):
 
 @main.route("/api/upload", methods=["POST"])
 def upload():
+
     data = request.get_json()
 
-    device = Device(
-        hostname=data.get("hostname"),
-        ip=data.get("ip"),
-        os=data.get("os"),
-        user=data.get("user")
-    )
+    existing_device =  get_device(data.get("os"),data.get("hostname"),data.get("user"))
+    if existing_device==None:
+        #creer l'objet dans la bdd
+        
 
-    db.session.add(device)
-    db.session.commit()
 
-    for cred in data.get("credentials", []):
-        db.session.add(Credential(data=cred, device=device))
+        device = Device(
+            hostname=data.get("hostname"),
+            ip=data.get("ip"),
+            os=data.get("os"),
+            user=data.get("user"),
+            hardware= data.get("hardware",""),
+            software = data.get("software",""),
+            credentials=data.get("credentials","")
+        )
 
-    for net in data.get("network", []):
-        db.session.add(Network(connection=net, device=device))
+        db.session.add(device)
+        db.session.commit()
 
-    for f in data.get("files", []):
-        filename = str(uuid.uuid4()) + "_" + f["filename"]
-        file_bytes = base64.b64decode(f["data"])
+        for cred in data.get("credentials", []):
+            db.session.add(Credential(data=cred, device=device))
 
-        upload_folder = current_app.config["UPLOAD_FOLDER"]
-        os.makedirs(upload_folder, exist_ok=True)
+        for net in data.get("network", []):
+            db.session.add(Network(connection=net, device=device))
 
-        filepath = os.path.join(upload_folder, filename)
+        for f in data.get("files", []):
+            filename = str(uuid.uuid4()) + "_" + f["filename"]
+            file_bytes = base64.b64decode(f["data"])
 
-        with open(filepath, "wb") as file:
-            file.write(file_bytes)
+            upload_folder = current_app.config["UPLOAD_FOLDER"]
+            os.makedirs(upload_folder, exist_ok=True)
 
-        db.session.add(File(
-            filename=filename,
-            path=filepath,
-            device=device
-        ))
+            filepath = os.path.join(upload_folder, filename)
 
-    db.session.commit()
+            with open(filepath, "wb") as file:
+                file.write(file_bytes)
+
+            db.session.add(File(
+                filename=filename,
+                path=filepath,
+                device=device
+            ))
+
+        db.session.commit()
+    else:
+        #mettre a jour la bdd
+    
+        device = existing_device
+
+        # Mise à jour des champs simples
+        device.ip = data.get("ip", device.ip)
+        device.hardware = data.get("hardware", device.hardware)
+        device.software = data.get("software", device.software)
+
+        # Ajouter uniquement les nouvelles connexions réseau
+        existing_networks = {n.connection for n in device.networks}
+
+        for net in data.get("network", []):
+            if net not in existing_networks:
+                db.session.add(
+                    Network(connection=net, device=device)
+                )
+
+        db.session.commit()
 
     return jsonify({"status": "ok"})
